make mvn clean install work. Commented IT without test spring context + some prittify
