+++ pom.xml
@@ -18,6 +18,8 @@
         <hibernate-validator.version>4.3.1.Final</hibernate-validator.version>
         <org.mapstruct.version>1.2.0.Beta1</org.mapstruct.version>
         <org.projectlombok.version>1.16.14</org.projectlombok.version>
+        <rest-assured.version>3.0.2</rest-assured.version>
+        <junit.version>4.12</junit.version>
         <hsqldb.version>2.3.2</hsqldb.version>
     </properties>
 
@@ -82,6 +84,7 @@
             <version>${org.projectlombok.version}</version>
             <scope>provided</scope>
         </dependency>
+        <!-- slf4j-log4j12 -->
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-log4j12</artifactId>
@@ -97,14 +100,14 @@
         <dependency>
             <groupId>io.rest-assured</groupId>
             <artifactId>rest-assured</artifactId>
-            <version>3.0.2</version>
+            <version>${rest-assured.version}</version>
             <scope>test</scope>
         </dependency>
         <!-- rest-assured-spring-mock-mvc -->
         <dependency>
             <groupId>io.rest-assured</groupId>
             <artifactId>spring-mock-mvc</artifactId>
-            <version>3.0.2</version>
+            <version>${rest-assured.version}</version>
             <scope>test</scope>
         </dependency>
         <!-- jsonsimple -->
@@ -125,7 +128,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.12</version>
+            <version>${junit.version}</version>
             <scope>test</scope>
         </dependency>
         <!-- mockito-all -->
@@ -153,17 +156,17 @@
             <artifactId>commons-lang3</artifactId>
             <version>3.0</version>
         </dependency>
-        <!-- HyperSQL DB -->
-        <dependency>
-            <groupId>org.hsqldb</groupId>
-            <artifactId>hsqldb</artifactId>
-            <version>${hsqldb.version}</version>
-        </dependency>
         <!-- PDFBox -->
         <dependency>
             <groupId>org.apache.pdfbox</groupId>
             <artifactId>pdfbox</artifactId>
             <version>2.0.4</version>
+        </dependency>
+        <!-- HyperSQL DB -->
+        <dependency>
+            <groupId>org.hsqldb</groupId>
+            <artifactId>hsqldb</artifactId>
+            <version>${hsqldb.version}</version>
         </dependency>
     </dependencies>
     <build>
