+++ src/main/java/com/opinta/dao/AddressDao.java
@@ -5,9 +5,14 @@ import java.util.List;
 import com.opinta.entity.Address;
 
 public interface AddressDao {
+
     List<Address> getAll();
+
     Address getById(long id);
+
     Address save(Address address);
+
     void update(Address address);
+
     void delete(Address address);
 }
