+++ src/main/java/com/opinta/dao/BarcodeInnerNumberDao.java
@@ -15,5 +15,4 @@ public interface BarcodeInnerNumberDao {
     void update(BarcodeInnerNumber barcodeInnerNumber);
     
     void delete(BarcodeInnerNumber barcodeInnerNumber);
-    
 }
