+++ src/main/java/com/opinta/dao/PostcodePoolDao.java
@@ -5,9 +5,14 @@ import java.util.List;
 import com.opinta.entity.PostcodePool;
 
 public interface PostcodePoolDao {
+
     List<PostcodePool> getAll();
+
     PostcodePool getById(long id);
+
     PostcodePool save(PostcodePool postcodePool);
+
     void update(PostcodePool postcodePool);
+
     void delete(PostcodePool postcodePool);
 }
