+++ src/main/java/com/opinta/dao/ShipmentDao.java
@@ -18,5 +18,4 @@ public interface ShipmentDao {
     void update(Shipment shipment);
 
     void delete(Shipment shipment);
-
 }
