+++ src/main/java/com/opinta/dao/ShipmentTrackingDetailDao.java
@@ -4,9 +4,14 @@ import com.opinta.entity.ShipmentTrackingDetail;
 import java.util.List;
 
 public interface ShipmentTrackingDetailDao {
+
     List<ShipmentTrackingDetail> getAll();
+
     ShipmentTrackingDetail getById(long id);
+
     ShipmentTrackingDetail save(ShipmentTrackingDetail shipmentTrackingDetail);
+
     void update(ShipmentTrackingDetail shipmentTrackingDetail);
+
     void delete(ShipmentTrackingDetail shipmentTrackingDetail);
 }
