+++ src/main/java/com/opinta/dao/TariffGridDao.java
@@ -16,7 +16,7 @@ public interface TariffGridDao {
 
     void delete(TariffGrid tariffGrid);
 
-    TariffGrid getPriceByDimension(float weight, float length, W2wVariation w2wVariation);
+    TariffGrid getByDimension(float weight, float length, W2wVariation w2wVariation);
 
     TariffGrid getLast(W2wVariation w2wVariation);
 }
