+++ src/main/java/com/opinta/service/PDFGeneratorService.java
@@ -5,5 +5,4 @@ public interface PDFGeneratorService {
     byte[] generateLabel(long id);
 
     byte[] generatePostpay(long id);
-
 }
