+++ src/main/java/com/opinta/service/ShipmentServiceImpl.java
@@ -1,5 +1,12 @@
 package com.opinta.service;
 
+import com.opinta.dao.TariffGridDao;
+import com.opinta.entity.Address;
+import com.opinta.entity.DeliveryType;
+import com.opinta.entity.TariffGrid;
+import com.opinta.entity.W2wVariation;
+import com.opinta.util.AddressUtil;
+import java.math.BigDecimal;
 import java.util.List;
 
 import javax.transaction.Transactional;
@@ -24,18 +31,17 @@ import static org.apache.commons.beanutils.BeanUtils.copyProperties;
 public class ShipmentServiceImpl implements ShipmentService {
     private final ShipmentDao shipmentDao;
     private final ClientDao clientDao;
+    private final TariffGridDao tariffGridDao;
     private final ShipmentMapper shipmentMapper;
     private final BarcodeInnerNumberService barcodeInnerNumberService;
-    private final ShipmentCalculation shipmentCalculation;
 
     @Autowired
-    public ShipmentServiceImpl(ShipmentDao shipmentDao, ClientDao clientDao, ShipmentMapper shipmentMapper,
-                               ShipmentCalculation shipmentCalculation,
-                               BarcodeInnerNumberService barcodeInnerNumberService) {
+    public ShipmentServiceImpl(ShipmentDao shipmentDao, ClientDao clientDao, TariffGridDao tariffGridDao,
+                               ShipmentMapper shipmentMapper, BarcodeInnerNumberService barcodeInnerNumberService) {
         this.shipmentDao = shipmentDao;
         this.clientDao = clientDao;
+        this.tariffGridDao = tariffGridDao;
         this.shipmentMapper = shipmentMapper;
-        this.shipmentCalculation = shipmentCalculation;
         this.barcodeInnerNumberService = barcodeInnerNumberService;
     }
 
@@ -79,7 +85,7 @@ public class ShipmentServiceImpl implements ShipmentService {
 
         shipment.setSender(clientDao.getById(shipment.getSender().getId()));
         shipment.setRecipient(clientDao.getById(shipment.getRecipient().getId()));
-        shipment.setPrice(shipmentCalculation.calculatePrice(shipment));
+        shipment.setPrice(calculatePrice(shipment));
 
         return shipmentMapper.toDto(shipmentDao.save(shipment));
     }
@@ -123,5 +129,46 @@ public class ShipmentServiceImpl implements ShipmentService {
     public Shipment getEntityById(long id) {
         log.info("Getting postcodePool by id {}", id);
         return shipmentDao.getById(id);
+    }
+
+    private BigDecimal calculatePrice(Shipment shipment) {
+        log.info("Calculating price for shipment {}", shipment);
+
+        Address senderAddress = shipment.getSender().getAddress();
+        Address recipientAddress = shipment.getRecipient().getAddress();
+        W2wVariation w2wVariation = W2wVariation.COUNTRY;
+        if (AddressUtil.isSameTown(senderAddress, recipientAddress)) {
+            w2wVariation = W2wVariation.TOWN;
+        } else if (AddressUtil.isSameRegion(senderAddress, recipientAddress)) {
+            w2wVariation = W2wVariation.REGION;
+        }
+
+        TariffGrid tariffGrid = tariffGridDao.getLast(w2wVariation);
+        if (shipment.getWeight() < tariffGrid.getWeight() &&
+                shipment.getLength() < tariffGrid.getLength()) {
+            tariffGrid = tariffGridDao.getByDimension(shipment.getWeight(), shipment.getLength(), w2wVariation);
+        }
+
+        log.info("TariffGrid for weight {} per length {} and type {}: {}",
+                shipment.getWeight(), shipment.getLength(), w2wVariation, tariffGrid);
+
+        if (tariffGrid == null) {
+            return BigDecimal.ZERO;
+        }
+
+        float price = tariffGrid.getPrice() + getSurcharges(shipment);
+
+        return new BigDecimal(Float.toString(price));
+    }
+
+    private float getSurcharges(Shipment shipment) {
+        float surcharges = 0;
+        if (shipment.getDeliveryType().equals(DeliveryType.D2W) ||
+                shipment.getDeliveryType().equals(DeliveryType.W2D)) {
+            surcharges += 9;
+        } else if (shipment.getDeliveryType().equals(DeliveryType.D2D)) {
+            surcharges += 12;
+        }
+        return surcharges;
     }
 }
